style: rename **/actions.ts from **/utils.ts
Rename `**/actions.ts` to distinguish it from `**/utils.ts`.
