+++ components/LoginForm/index.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import Link from "next/link";
 import UnPwInputs from "@/types/UnPwInputs";
 import UnPwForm from "../UnPwForm";
-import { login } from "./utils";
+import { login } from "./actions";
 import style from "./style.module.scss";
 
 interface LoginFormProps {
